gql-schema-dumper/pyproject.toml
[tool.poetry]
name = "gql-schema-dumper"
version = "0.1.0"
description = ""
authors = ["Shogo Sawai <shogo.sawai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
graphql-core = "^3.2.3"
watchfiles = "^0.18.1"
requests = "^2.28.1"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.20.0"
radon = "^5.1.0"
pyproject-flake8 = "^5.0.4.post1"
mypy = "^0.990"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.isort]
line_length = 110
profile = "black"

[tool.flake8]
max-line-length = 110
max-complexity = 18
extend-ignore = "E203,"  # blackと競合してる
